new task

+++ ProgramingFundamentalsJava/DataTypesandVariables_Lab/PoundstoDollars_02.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class PoundstoDollars_02 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int pound = Integer.parseInt(scanner.nextLine());
+        double pound = Double.parseDouble(scanner.nextLine());
         double dollars = pound * 1.36;
         System.out.printf("%.3f",dollars);
     }
